refactor: separate steps into their own components
